WIP: refactor - wrap CodeMirrorEditor with Editor
